Working on Test for Center

+++ ticker_gtk.py
@@ -53,7 +53,7 @@ def static_ticker_center():
     # set the pygame window name
     pygame.display.set_caption('StaticTicker')
 
-    font_size_for_center = conf['static_ticker_font_size'] / 2
+    font_size_for_center = int(conf['static_ticker_font_size'] / 2)
     print(font_size_for_center)
 
     fonting = pygame.font.SysFont(conf['static_ticker_font'], 40)
